@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy, ViewChild, ComponentFactoryResolver } from '@angular/core';
 import {Router, ActivatedRoute} from '@angular/router';
 
-import { Subscription }   from 'rxjs/Subscription';
+import { Subscription }   from 'rxjs';
 
 import {UserService} from '../../shared/services/user.service';
 import {EventAnalysisService} from '../../shared/services/event-analysis.service';
